requirements.txt
alembic==1.14.1
blinker==1.8.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cors==1.0.1
cryptography==45.0.6
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==24.2.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
psycopg2==2.9.10
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.4
requests-file==2.1.0
SQLAlchemy==2.0.43
tldextract==5.1.2
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
zope.event==5.0
zope.interface==7.2
